zeraFa: split icon() into colorize() + icon()
The function icon() does two things at one time: colorize glyph and append
space for a margin. There are cases we just want colored glyphs so introduce
colorize().

This was stolen from fontawesome-qml [1] and in the long run we should replace
zerafa by fontawesome-qml (as discussed) but now is not the time...

[1] https://github.com/schnitzeltony/fontawesome-qml

Signed-off-by: Andreas Müller <schnitzeltony@gmail.com>