uivectorgraphics: Replace font metrics by approximation / revamp label positions
During test creation we learned that font-metrics adds us unreproducible label
positions: Font found in the test environment are a varaible not under our
control.
With the test-dot introduced for manual tests we found that the optical
accuracy for this solution is pretty good even on long labels found in 3-wire
mode.

Signed-off-by: Andreas Müller <schnitzeltony@gmail.com>